New metrics service
Introduce a new metrics service as a preliminary step for having new metrics stats
Some refactoring of the existing:
- Move some istio-related functions from prometheus client to this new metrics service (keep client more low level)
- There was already a metrics handler file, but only used for Aggregation-kind metrics. Unify it to include app metrics, service metrics and workload metrics handlers. It will also hold the new metrics stats handler.
- Create metrics model file (high level model for REST API was previously in prometheus client package, having the new business layer allows to decouple client // model)
- Move existing tests to the relevant package as well
