+++ business/apps.go
@@ -122,8 +122,7 @@ func (in *AppService) GetApp(namespace string, appName string) (models.App, erro
 	for _, workload := range appDetails.Workloads {
 		pods = append(pods, workload.Pods...)
 	}
-	dash := NewDashboardsService(in.prom)
-	(*appInstance).Runtimes = dash.GetCustomDashboardRefs(namespace, appName, "", pods)
+	(*appInstance).Runtimes = NewDashboardsService().GetCustomDashboardRefs(namespace, appName, "", pods)
 
 	return *appInstance, nil
 }
