+++ handlers/metrics_test.go
@@ -44,7 +44,7 @@ func TestExtractMetricsQueryParams(t *testing.T) {
 	q.Add("requestProtocol", "http")
 	req.URL.RawQuery = q.Encode()
 
-	mq := prometheus.IstioMetricsQuery{Namespace: "ns"}
+	mq := business.IstioMetricsQuery{Namespace: "ns"}
 	err = extractIstioMetricsQueryParams(req, &mq, buildNamespace("ns", time.Time{}))
 	if err != nil {
 		t.Fatal(err)
@@ -78,7 +78,7 @@ func TestExtractMetricsQueryParamsStepLimitCase(t *testing.T) {
 	q.Add("duration", "1000")        // Makes start = 2018-04-10T12:24:20
 	req.URL.RawQuery = q.Encode()
 
-	mq := prometheus.IstioMetricsQuery{Namespace: "ns"}
+	mq := business.IstioMetricsQuery{Namespace: "ns"}
 	err = extractIstioMetricsQueryParams(req, &mq, buildNamespace("ns", time.Time{}))
 	if err != nil {
 		t.Fatal(err)
@@ -103,7 +103,7 @@ func TestExtractMetricsQueryIntervalBoundary(t *testing.T) {
 	q.Add("rateInterval", "35m")
 	req.URL.RawQuery = q.Encode()
 
-	mq := prometheus.IstioMetricsQuery{Namespace: "ns"}
+	mq := business.IstioMetricsQuery{Namespace: "ns"}
 	err = extractIstioMetricsQueryParams(req, &mq, buildNamespace("ns", time.Date(2018, 4, 10, 12, 10, 0, 0, time.UTC)))
 	if err != nil {
 		t.Fatal(err)
@@ -125,7 +125,7 @@ func TestExtractMetricsQueryStartTimeBoundary(t *testing.T) {
 	q.Add("rateInterval", "1m")
 	req.URL.RawQuery = q.Encode()
 
-	mq := prometheus.IstioMetricsQuery{Namespace: "ns"}
+	mq := business.IstioMetricsQuery{Namespace: "ns"}
 	namespaceTimestamp := time.Date(2018, 4, 10, 12, 30, 0, 0, time.UTC)
 
 	err = extractIstioMetricsQueryParams(req, &mq, buildNamespace("ns", namespaceTimestamp))
