+++ prometheus/types.go
@@ -7,8 +7,8 @@ import (
 	"github.com/prometheus/common/model"
 )
 
-// BaseMetricsQuery holds common parameters for all kinds of queries
-type BaseMetricsQuery struct {
+// RangeQuery holds common parameters for all kinds of range queries
+type RangeQuery struct {
 	prom_v1.Range
 	RateInterval string
 	RateFunc     string
@@ -18,56 +18,13 @@ type BaseMetricsQuery struct {
 }
 
 // FillDefaults fills the struct with default parameters
-func (q *BaseMetricsQuery) fillDefaults() {
+func (q *RangeQuery) FillDefaults() {
 	q.End = time.Now()
 	q.Start = q.End.Add(-30 * time.Minute)
 	q.Step = 15 * time.Second
 	q.RateInterval = "1m"
 	q.RateFunc = "rate"
 	q.Avg = true
-}
-
-// IstioMetricsQuery holds query parameters for a typical metrics query
-type IstioMetricsQuery struct {
-	BaseMetricsQuery
-	Filters         []string
-	Namespace       string
-	App             string
-	Workload        string
-	Service         string
-	Direction       string // outbound | inbound
-	RequestProtocol string // e.g. http | grpc
-	Reporter        string // source | destination, defaults to source if not provided
-	Aggregate       string
-	AggregateValue  string
-}
-
-// FillDefaults fills the struct with default parameters
-func (q *IstioMetricsQuery) FillDefaults() {
-	q.BaseMetricsQuery.fillDefaults()
-	q.Reporter = "source"
-	q.Direction = "outbound"
-}
-
-// CustomMetricsQuery holds query parameters for a custom metrics query
-type CustomMetricsQuery struct {
-	BaseMetricsQuery
-	Namespace         string
-	App               string
-	Version           string
-	RawDataAggregator string
-}
-
-// FillDefaults fills the struct with default parameters
-func (q *CustomMetricsQuery) FillDefaults() {
-	q.BaseMetricsQuery.fillDefaults()
-	q.RawDataAggregator = "sum"
-}
-
-// Metrics contains all simple metrics and histograms data
-type Metrics struct {
-	Metrics    map[string]*Metric   `json:"metrics"`
-	Histograms map[string]Histogram `json:"histograms"`
 }
 
 // Metric holds the Prometheus Matrix model, which contains one or more time series (depending on grouping)
